requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
google-generativeai
python-multipart
pandas
prophet
ortools
redis
celery
pydantic
python-jose[cryptography]
passlib[bcrypt]
pillow